Adding 6 more pattern test cases to patterns_general.json
These additions have the potential to break naïve replacement solutions. Since ?, * and [ are
not special in glob character classes, they do not need to be escaped. Naïve replacement
solutions are unlikely to take this into account. For example:

"[*]" should only match "*"

A simple function that translates a glob pattern to regex may naïvly replace "*" → ".*". As
such, the regular expression check will be against "[.*]", which now matches 2 characters.
Note that whist the example is trivial, and it's unlikely someone would use "[*]" instead of
"\*", it's still broken. Further examples (such as "[*-/]") are entirely valid, and should
be available for use.